Added styles and skills controllers

+++ src/main/java/com/wearperfect/dataservice/api/dto/UserSkillDTO.java
@@ -9,7 +9,7 @@ public class UserSkillDTO {
 	
 	Long userId;
 	
-	Long skillId;
+	Integer skillId;
 	
 	Long createdBy;
